Helper functions transferred from interpretation.ipynb to src/interpretation
