+++ src/configs/interpretation_config.yaml
@@ -10,7 +10,7 @@ model:
     return_attention: true
 
 base:
-    project_ids: ["GBM", "LGG"]
+    project_ids: ["BRCA"] #["GBM", "LGG"]
     device: 'cuda:3'
     remove_nan_column: 'WSI_initial'
     batch_size: 8 #to compute stattest on larger amount of losses
